feat: move ignore option tests
